create trace.py
